docker/SPAM/python/requirements.txt
tensorflow-cpu==2.19.0
numpy
flask